(hybris) ignore "mount" and "mkdir /tmp" commands entirely.
mount is usually removed during build, but if the *.rc files are on
partitions we cannot easily change this is required (e.g. system,
vendor, odm, ...).

mkdir /tmp has the side effect that it changes the permissions on
/tmp. systemd owned directories shouldn't be changed by
droid-hal-init. In the future we might need to add other
directories to this blacklist as well.

unmount doens't make sense if we don't allow mount.

Change-Id: Iba637613c6a927098c947536e11348e269e1bef3